Implemented get SMS, and delete SMS to ensure all SMS messages are tested/avoid repeats
The app now, on pressing 'GO', gets the latest SMS, test the website
gotten from the SMS, then deletes the SMS from the Twilio account. It is
a made queuing system of some sort to make sure each URL texted to us is
tested. Will later implement this as a button press on the raspberry pi.
The SMS messages are read FIFO